+++ AppboyProject/index.ios.js
@@ -1,8 +1,6 @@
 'use strict';
-import React, {
-  AppRegistry,
-  Component,
-} from 'react-native';
-import AppboyProject from './AppboyProject.js'
+import React, { Component } from 'react';
+import { AppRegistry } from 'react-native';
+import AppboyProject from './AppboyProject.js';
 
 AppRegistry.registerComponent('AppboyProject', () => AppboyProject);