+++ AppboyProject/ios/AppboyProject/ABKReactBridgeTest.h
@@ -1,6 +1,0 @@
-#import <Foundation/Foundation.h>
-#import "RCTBridgeModule.h"
-
-@interface ABKReactBridgeTest : NSObject <RCTBridgeModule>
-
-@end